Split ConfirmSeedActivity into VerifySeedActivity & ConfirmSeedActivity
