+++ app/src/main/java/com/dcrandroid/activities/ConfirmSeedActivity.kt
@@ -19,37 +19,28 @@ import android.widget.RelativeLayout
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.dcrandroid.R
-import com.dcrandroid.adapter.CreateWalletAdapter
 import com.dcrandroid.adapter.InputSeed
-import com.dcrandroid.adapter.MultiSeed
 import com.dcrandroid.adapter.RestoreWalletAdapter
 import com.dcrandroid.data.Constants
 import dcrlibwallet.Dcrlibwallet
 import kotlinx.android.synthetic.main.confirm_seed_page.*
-import java.util.*
-import kotlin.collections.ArrayList
-
 
 private const val CLICK_THRESHOLD = 300 //millisecond
 
-// TODO: Split this class into VerifySeedActivity  for old wallet and EnterSeedActivity for new wallet.
 class ConfirmSeedActivity : BaseActivity(), View.OnTouchListener {
 
-    private var restore: Boolean = false
     private var verifiedSeed: Boolean = false
+    private var lastConfirmClick: Long = 0
+    private var finalSeedsString = ""
+
     private var allSeeds = ArrayList<String>()
     private val seedsForInput = ArrayList<InputSeed>()
     private val confirmedSeedsArray = ArrayList<InputSeed>()
     private var sortedList = listOf<InputSeed>()
-    private val arrayOfRandomSeeds = ArrayList<InputSeed>()
-    private val arrayOfSeedLists = ArrayList<MultiSeed>()
-    private val shuffledSeeds = ArrayList<InputSeed>()
-    private var finalSeedsString = ""
-    private var lastConfirmClick: Long = 0
+
     private var handler: Handler? = null
-    private var currentSeedPosition = 0
+
     private lateinit var restoreWalletAdapter: RestoreWalletAdapter
-    private lateinit var createWalletAdapter: CreateWalletAdapter
     private lateinit var linearLayoutManager: LinearLayoutManager
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -67,25 +58,17 @@ class ConfirmSeedActivity : BaseActivity(), View.OnTouchListener {
     private fun prepareData() {
         val bundle = intent.extras
         val temp = arrayOfNulls<String>(33)
-        if (!bundle.isEmpty) {
+        if (bundle != null && !bundle.isEmpty) {
             val seed = bundle.getString(Constants.SEED)
-            restore = bundle.getBoolean(Constants.RESTORE)
             allSeeds = ArrayList(seed.split(" "))
-            if (restore) {
-                temp.forEachIndexed { number, _ ->
-                    seedsForInput.add(InputSeed(number, " "))
-                    initOldWalletAdapter()
-                }
-            } else {
-                tvHeader.text = getString(R.string.seed_phrase_verification)
-                tvHint.text = getString(R.string.please_confirm_your_seed_by_typing_and_tapping_each_word_accordingly)
-                headerTop.isFocusableInTouchMode = true
-                initNewWalletAdapter()
+            temp.forEachIndexed { number, _ ->
+                seedsForInput.add(InputSeed(number, " "))
+                initSeedAdapter()
             }
         }
     }
 
-    private fun initOldWalletAdapter() {
+    private fun initSeedAdapter() {
         restoreWalletAdapter = RestoreWalletAdapter(seedsForInput, allSeeds, applicationContext,
                 { savedSeed: InputSeed ->
                     confirmedSeedsArray.add(savedSeed)
@@ -118,81 +101,6 @@ class ConfirmSeedActivity : BaseActivity(), View.OnTouchListener {
             }
         })
         recyclerViewSeeds.adapter = restoreWalletAdapter
-    }
-
-    private fun initNewWalletAdapter() {
-        createWalletAdapter = CreateWalletAdapter(applicationContext, arrayOfSeedLists, { enteredSeeds: ArrayList<InputSeed> ->
-            confirmedSeedsArray.clear()
-            confirmedSeedsArray.addAll(enteredSeeds)
-            sortedList = confirmedSeedsArray.sortedWith(compareBy { it.number }).distinct()
-
-            if (sortedList.size < 33) {
-                tvError.visibility = View.VISIBLE
-                tvError.text = getString(R.string.notAllSeedsEntered)
-            }
-            val itemView = recyclerViewSeeds.findViewById<RelativeLayout>(R.id.rlButtons)
-            val itemHeight = itemView.measuredHeight
-            val maxAllowedHeight = nestedScrollView.getChildAt(0).bottom - (itemView.measuredHeight + itemView.measuredHeight / 2)
-
-            val currentHeight = (nestedScrollView.scrollY + nestedScrollView.height)
-            if (currentHeight > maxAllowedHeight) {
-                nestedScrollView.smoothScrollBy(0, itemHeight * 2)
-            }
-        }, { isAllEntered: Boolean ->
-            if (isAllEntered && sortedList.size == 33) {
-                handleSingleTap()
-            }
-        })
-        recyclerViewSeeds.adapter = createWalletAdapter
-        generateRandomSeeds()
-    }
-
-    private fun generateRandomSeeds() {
-        val firstRandom = (1..allSeeds.size).random()
-        val secondRandom = (1..allSeeds.size).random()
-        var currentItemPosition = 0
-        if (currentSeedPosition != 33) {
-            currentItemPosition = (currentSeedPosition)
-        }
-
-        if (firstRandom != secondRandom && firstRandom != currentItemPosition && secondRandom != currentItemPosition) {
-            for (item in allSeeds) {
-                when (item) {
-                    allSeeds[firstRandom] -> arrayOfRandomSeeds.add(InputSeed(firstRandom, item))
-                    allSeeds[secondRandom] -> arrayOfRandomSeeds.add(InputSeed(secondRandom, item))
-                    allSeeds[currentItemPosition] -> arrayOfRandomSeeds.add(InputSeed(currentItemPosition, item))
-                }
-            }
-            addSeedsToAdapter()
-        } else {
-            generateRandomSeeds()
-        }
-    }
-
-    private fun addSeedsToAdapter() {
-        shuffledSeeds.addAll(arrayOfRandomSeeds.shuffled().distinct())
-        if (shuffledSeeds.size == 3) {
-            when {
-                currentSeedPosition < 32 -> {
-                    arrayOfSeedLists.add(MultiSeed(shuffledSeeds[0], shuffledSeeds[1], shuffledSeeds[2]))
-                    createWalletAdapter.notifyDataSetChanged()
-                    currentSeedPosition++
-                    arrayOfRandomSeeds.clear()
-                    shuffledSeeds.clear()
-                    generateRandomSeeds()
-                }
-                currentSeedPosition == 32 -> {
-                    arrayOfSeedLists.add(MultiSeed(shuffledSeeds[0], shuffledSeeds[1], shuffledSeeds[2]))
-                    createWalletAdapter.notifyDataSetChanged()
-                    arrayOfRandomSeeds.clear()
-                    shuffledSeeds.clear()
-                }
-            }
-        } else {
-            arrayOfRandomSeeds.clear()
-            shuffledSeeds.clear()
-            generateRandomSeeds()
-        }
     }
 
     override fun onTouch(v: View?, event: MotionEvent?): Boolean {
@@ -249,12 +157,7 @@ class ConfirmSeedActivity : BaseActivity(), View.OnTouchListener {
                 }
 
                 tvError.visibility = View.VISIBLE
-
-                if (restore) {
-                    tvError.text = getString(R.string.restore_wallet_incorrect_seed_input)
-                } else {
-                    tvError.text = getString(R.string.create_wallet_incorrect_seeds_input)
-                }
+                tvError.text = getString(R.string.restore_wallet_incorrect_seed_input)
             }
 
         } else if (sortedList.isNotEmpty() && (sortedList.size != 33)) {
@@ -263,6 +166,4 @@ class ConfirmSeedActivity : BaseActivity(), View.OnTouchListener {
             tvError.text = getString(R.string.theInputFieldIsEmpty)
         }
     }
-
-    private fun IntRange.random() = Random().nextInt((allSeeds.size + 1) - start)
 }
